+++ tests/testthat/test_get_data.R
@@ -22,24 +22,27 @@ test_that("nomis_get_data return expected format", {
   expect_type(a, "list")
   expect_true(tibble::is_tibble(a))
   expect_equal(
-    as.numeric(a$RECORD_OFFSET[[nrow(a)]]) + 1,
-    as.numeric(a$RECORD_COUNT[[nrow(a)]])
+    as.numeric(a$record_offset[[nrow(a)]]) + 1,
+    as.numeric(a$record_count[[nrow(a)]])
   )
-  sum_check <- summary(diff(as.numeric(a$RECORD_OFFSET)))
+  sum_check <- summary(diff(as.numeric(a$record_offset)))
   expect_equal(sum_check[[1]], 1)
 
   expect_error(nomis_get_data(), "Dataset ID must be specified")
 
-  b <- nomis_get_data(
-    id = "NM_168_1", time = "latest",
-    geography = "2013265925", sex = "0"
+  expect_message(
+    b <- nomis_get_data(
+      id = "NM_168_1", time = "latest",
+      geography = "2013265925", sex = "0", additional_queries = ""
+    ), "The `additional_query` parameter is
+            deprecated, please use ... instead"
   )
 
-  expect_true(nrow(b) == b$RECORD_COUNT[1])
+  expect_true(nrow(b) == b$record_count[1])
   expect_length(b, 40)
   expect_type(b, "list")
   expect_true(tibble::is_tibble(b))
-  expect_true(nrow(a) == a$RECORD_COUNT[1])
+  expect_true(nrow(a) == a$record_count[1])
 
   c <- nomis_get_data(
     id = "NM_127_1", sex = "6",
@@ -48,7 +51,7 @@ test_that("nomis_get_data return expected format", {
   expect_length(c, 28)
   expect_type(c, "list")
   expect_true(tibble::is_tibble(c))
-  expect_true(nrow(c) == c$RECORD_COUNT[1])
+  expect_true(nrow(c) == c$record_count[1])
 
   expect_error(nomis_get_data(
     id = "NM_1_1", time = "latest",
@@ -68,15 +71,30 @@ test_that("nomis_get_data return expected format", {
   expect_type(x_select, "list")
   expect_true(tibble::is_tibble(x_select))
 
-  mort_data1 <- nomis_get_data(id = "NM_161_1", date = "2016",
-                              geography = "TYPE464", 
-                              CAUSE_OF_DEATH = "10381", 
-                              sex = 0, age = 0, MEASURE = 6)
-  
-  mort_data2 <- nomis_get_data(id = "NM_161_1", date = "2016",
-                               geography = "TYPE464", sex = 0, measures = NULL,
-                               cause_of_death = "10381", 
-                               age = 0, measure = "6")
-  
+  mort_data1 <- nomis_get_data(
+    id = "NM_161_1", date = "2016",
+    geography = "TYPE464",
+    CAUSE_OF_DEATH = "10381",
+    sex = 0, age = 0, MEASURE = 6
+  )
+
+  mort_data2 <- nomis_get_data(
+    id = "NM_161_1", date = "2016",
+    geography = "TYPE464", sex = 0,
+    cause_of_death = "10381",
+    age = 0, measure = 6, measures = NULL
+  )
+
   expect_true(all.equal(mort_data2, mort_data1))
+  expect_true(is.numeric(mort_data2$obs_value))
+
+  expect_error(
+    mort_data3 <- nomis_get_data(
+      id = "NM_161_1", date = "2016",
+      geography = "TYPE464", sex = 0,
+      cause_of_death = "10381",
+      age = 0, measure = 6
+    ),
+    "The API request did not return any results.\nPlease check your parameters."
+  )
 })